requirements.txt
open_clip_torch
huggingface_hub
Pillow
torch 
torchvision
torchaudio
timm
argparse
transformer